Fix nucleotide represntation bias in PickIlluminaIndices.
The kmers where being generated in lexicographical order, and so various
filters (ex. reverese and palindrom) may preferentially keep the kmers
that were generated eaerlier, and thus keep kmers that are smaller
lexicographically (i.e. start with A).  We shuffle prior to applying
filters that may have such a bias.