Fixed nasty bug in the cumulative error plot for ErrorRateByReadPosition.
There's a fairly well documented failure when using ggplot with aes(color=some_var, y=cumsum(other_var)).  The expectation is that you'd get the cumulative sum over the partition of the data for the appropriate color, but what actually happens is that the cumsum is evaluated on the entire dataframe, then it's plotted by color.  So later "color"s start out, not at zero, but at the previous total.  In our case the R2 error sum was starting at the final sum of the R1 errors!