[feature] Reduce memory usage by GroupReadsByUmi when not allowing diffs between UMIs (#774)
Makes a change to GroupReadsByUmi to extend how it sorts reads to include the raw UMI sequence in cases where no differences are allowed between UMIs in grouping, then leverages that information to draw reads in in small chunks for grouping.  This can cause a dramatic reduction in memory usage when used with data that has a high frequency of the same start/stop position of reads (e.g. multiplex PCR) _and_ either `--edits 0` is specified or `--strategy Identity`.

Introduces a small functional change, only in the case where a) no diffs are allowed and b) variable length UMIs are used with `--min-umi-length`.  In this case the tool will occasionally do less trimming of UMIs than it did previously, and therefore possibly create incrementally more groups.