Add the --annotate-all option to AssignPrimers (#669)
* Add the --annotate-all option to AssignPrimers

In cases where a given read end (eg. R2) has both a primary and
supplementary alignment, where only of the two is assigned a primer,
the current behavior is to annotate the mate primer/amplicon for both,
but the primer/amplicon just for the alignment for the read end that
is assigned a primer.

The --annotate-all optiona will annotate both alignments of the same end
with the primer.  This may be important for downstream translocation
detection.